requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
ollama==0.4.7
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
